Moved plotting function to its own moduel
